finalizing Home page -- need to move NavBar and to change font
