+++ src/App.css
@@ -13,7 +13,7 @@
   }
 }
 
-.App-header {
+/*.App-header {
   background-color: #f5ffff;
   min-height: 100vh;
   display: flex;
@@ -22,7 +22,7 @@
   justify-content: center;
   font-size: calc(10px + 2vmin);
   color: black;
-}
+}*/
 
 
 .navbar {
